Rename `IsEqual()` to `IsHandlerEqual()`.
