+++ entity_test.go
@@ -3,9 +3,7 @@ package configkit_test
 import (
 	. "github.com/dogmatiq/configkit"
 	cfixtures "github.com/dogmatiq/configkit/fixtures" // can't dot-import due to conflicts
-	"github.com/dogmatiq/configkit/message"
-	"github.com/dogmatiq/dogma"
-	"github.com/dogmatiq/dogma/fixtures" // can't dot-import due to conflicts
+	"github.com/dogmatiq/configkit/message"            // can't dot-import due to conflicts
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
@@ -211,84 +209,4 @@ var _ = Describe("type EntityMessageTypes", func() {
 			),
 		)
 	})
-})
-
-var _ = Describe("func IsEqual()", func() {
-	It("returns true if the two entities are equivalent", func() {
-		h := &fixtures.AggregateMessageHandler{
-			ConfigureFunc: func(c dogma.AggregateConfigurer) {
-				c.Identity("<name>", "<key>")
-				c.ConsumesCommandType(fixtures.MessageA{})
-				c.ConsumesCommandType(fixtures.MessageB{})
-				c.ProducesEventType(fixtures.MessageE{})
-			},
-		}
-
-		a := FromAggregate(h)
-		b := FromAggregate(h)
-
-		Expect(IsEqual(a, b)).To(BeTrue())
-	})
-
-	DescribeTable(
-		"returns false if the entities are not equivalent",
-		func(b Entity) {
-			h := &fixtures.AggregateMessageHandler{
-				ConfigureFunc: func(c dogma.AggregateConfigurer) {
-					c.Identity("<name>", "<key>")
-					c.ConsumesCommandType(fixtures.MessageA{})
-					c.ConsumesCommandType(fixtures.MessageB{})
-					c.ProducesEventType(fixtures.MessageE{})
-				},
-			}
-
-			a := FromAggregate(h)
-
-			Expect(IsEqual(a, b)).To(BeFalse())
-		},
-		Entry(
-			"type differs",
-			FromIntegration(&fixtures.IntegrationMessageHandler{
-				ConfigureFunc: func(c dogma.IntegrationConfigurer) {
-					c.Identity("<name>", "<key>")
-					c.ConsumesCommandType(fixtures.MessageA{})
-					c.ProducesEventType(fixtures.MessageB{}) // diff
-					c.ProducesEventType(fixtures.MessageE{})
-				},
-			}),
-		),
-		Entry(
-			"identity name differs",
-			FromAggregate(&fixtures.AggregateMessageHandler{
-				ConfigureFunc: func(c dogma.AggregateConfigurer) {
-					c.Identity("<name-different>", "<key>") // diff
-					c.ConsumesCommandType(fixtures.MessageA{})
-					c.ConsumesCommandType(fixtures.MessageB{})
-					c.ProducesEventType(fixtures.MessageE{})
-				},
-			}),
-		),
-		Entry(
-			"identity key differs",
-			FromAggregate(&fixtures.AggregateMessageHandler{
-				ConfigureFunc: func(c dogma.AggregateConfigurer) {
-					c.Identity("<name>", "<key-different>") // diff
-					c.ConsumesCommandType(fixtures.MessageA{})
-					c.ConsumesCommandType(fixtures.MessageB{})
-					c.ProducesEventType(fixtures.MessageE{})
-				},
-			}),
-		),
-		Entry(
-			"messages differ",
-			FromAggregate(&fixtures.AggregateMessageHandler{
-				ConfigureFunc: func(c dogma.AggregateConfigurer) {
-					c.Identity("<name>", "<key>")
-					c.ConsumesCommandType(fixtures.MessageA{})
-					c.ProducesEventType(fixtures.MessageB{}) // diff
-					c.ProducesEventType(fixtures.MessageE{})
-				},
-			}),
-		),
-	)
 })
